add Admin auth

+++ server/src/routes/gymclasses.js
@@ -13,44 +13,54 @@ router.get('/', (req, res) => {
     .catch(err => res.status(500).json({ errors: { global: "Something went wrong" }}));
 });
 
-router.get('/:_id', (req, res) => {
+router.get('/:_id', authenticate, (req, res) => {
   GymClass.findById(req.params._id)
     .then(gymclass => res.json({ gymclass }))
     .catch(err => res.status(500).json({ errors: { global: "Something went wrong" }}));
 });
 
 router.delete('/:_id', authenticate, (req, res) => {
-  console.log(req.currentUser);
-  GymClass.findByIdAndRemove(req.params._id)
+  if (req.currentUser.username === 'admin'){
+    GymClass.findByIdAndRemove(req.params._id)
     .then(gymclass => res.json({ gymclass }))
     .catch(err => res.status(500).json({ errors: { global: "Something went wrong" }}));
+  } else {
+    res.status(401).json({ error: 'Failed to authenticate' });
+  }
 });
 
-
 router.post('/', authenticate, (req, res) => {
-  let { errors, isValid } = validateInput(req.body);
+  if (req.currentUser.username === 'admin'){
+    let { errors, isValid } = validateInput(req.body);
 
-  if (isValid) {
-    const { title, date } = req.body;
-    GymClass.create({ title, date })
+    if (isValid) {
+      const { title, date } = req.body;
+      GymClass.create({ title, date })
       .then(event => res.json({ gymclass: event }))
       .catch(err => res.status(500).json({ errors: { global: "Something went wrong" }}));
+    } else {
+      res.status(400).json({ errors });
+    }
   } else {
-    res.status(400).json({ errors });
+    res.status(401).json({ errors: { global: 'Failed to authenticate' } });
   }
 });
 
 router.put('/:_id', authenticate, (req, res) => {
   console.log(req.currentUser);
-  let { errors, isValid } = validateInput(req.body);
+  if (req.currentUser.username === 'admin'){
+    let { errors, isValid } = validateInput(req.body);
 
-  if (isValid) {
-    const { title, date } = req.body;
-    GymClass.findByIdAndUpdate(req.params._id, { title, date }, { new: true })
+    if (isValid) {
+      const { title, date } = req.body;
+      GymClass.findByIdAndUpdate(req.params._id, { title, date }, { new: true })
       .then(gymclass => res.json({ gymclass: gymclass }))
       .catch(err => res.status(500).json({ errors: { global: "Something went wrong" }}));
+    } else {
+      res.status(400).json({ errors });
+    }
   } else {
-    res.status(400).json({ errors });
+    res.status(401).json({ errors: { global: 'Failed to authenticate' } });
   }
 });
